feat: add storage providers which can generate s3 buckets and keys
