+++ pangea/core/models/__init__.py
@@ -1,6 +1,6 @@
 
 from .organization import Organization, PangeaUser
-from .s3 import S3ApiKey, S3Bucket
+from .s3 import S3ApiKey, S3Bucket, S3Provider
 from .sample_group import SampleGroup, SampleLibrary
 from .sample import Sample
 from .analysis_result import (
